refactor: file-manager naming
